mv extractObjectSchema() to io-schema.ts.
