+++ express-zod-api/tests/documentation-helpers.spec.ts
@@ -45,7 +45,6 @@ import {
   ensureShortDescription,
   excludeExamplesFromDepiction,
   excludeParamsFromDepiction,
-  extractObjectSchema,
   defaultIsHeader,
   onEach,
   onMissing,
@@ -84,56 +83,6 @@ describe("Documentation helpers", () => {
 
   beforeEach(() => {
     makeRefMock.mockClear();
-  });
-
-  describe("extractObjectSchema()", () => {
-    test("should pass the object schema through", () => {
-      const subject = extractObjectSchema(z.object({ one: z.string() }));
-      expect(subject).toBeInstanceOf(z.ZodObject);
-      expect(subject).toMatchSnapshot();
-    });
-
-    test("should return object schema for the union of object schemas", () => {
-      const subject = extractObjectSchema(
-        z.object({ one: z.string() }).or(z.object({ two: z.number() })),
-      );
-      expect(subject).toBeInstanceOf(z.ZodObject);
-      expect(subject).toMatchSnapshot();
-    });
-
-    test("should return object schema for the intersection of object schemas", () => {
-      const subject = extractObjectSchema(
-        z.object({ one: z.string() }).and(z.object({ two: z.number() })),
-      );
-      expect(subject).toBeInstanceOf(z.ZodObject);
-      expect(subject).toMatchSnapshot();
-    });
-
-    test("should support ez.raw()", () => {
-      const subject = extractObjectSchema(ez.raw());
-      expect(subject).toBeInstanceOf(z.ZodObject);
-      expect(subject).toMatchSnapshot();
-    });
-
-    describe("Feature #600: Top level refinements", () => {
-      test("should handle refined object schema", () => {
-        const subject = extractObjectSchema(
-          z.object({ one: z.string() }).refine(() => true),
-        );
-        expect(subject).toBeInstanceOf(z.ZodObject);
-        expect(subject).toMatchSnapshot();
-      });
-    });
-
-    describe("Feature #1869: Top level transformations", () => {
-      test("should handle transformations to another object", () => {
-        const subject = extractObjectSchema(
-          z.object({ one: z.string() }).transform(({ one }) => ({ two: one })),
-        );
-        expect(subject).toBeInstanceOf(z.ZodObject);
-        expect(subject).toMatchSnapshot();
-      });
-    });
   });
 
   describe("excludeParamsFromDepiction()", () => {
